Moved blog-header to ui folder
