+++ blogposts/00005.utf8clientbug.tsx
@@ -1,6 +1,6 @@
 import { CodeBlock } from "@/components/ui/codeblock";
 import { Metadata } from "@/lib/blogType";
-import BlogHeader from "@/components/blog-header";
+import BlogHeader from "@/components/ui/blog-header";
 
 export const metadata: Metadata = {
   title:
